resolve review comment, refactor some methods
